backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.9.0
pydantic==2.5.0
openai==1.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
numpy==2.0.2
# ソロモード用の追加ライブラリ
torch==2.1.0
torchaudio==2.1.0
transformers==4.36.0